Fix PWA always logging out when launched anew
Seems to be due to its inital route being dictated by the manifest.json,
thus resulting in '/index.html' instead of some '/:key' route. We instead
realy on remembering collective data from localStorage and redirect if
there is nothing to remember.

This fix is kind of experimental, but let's give it a try.