Remove uuid check from url and rely on localStorage instead
This simplifies the whole affair: When an invalud URL is loaded, we will
fall back to the last remembered Collective we've been visiting. This will
hopefully be more intutitive for any PWA, as it will just load the previous
view. We can still logout and visit a different URL to swap to a new, yet
unvisited Collective, otherwise we'll be able to pick it from the /unknwown
route.